Convert Plugin Module to TS
