--- src/plugin/index.js
@@ -1,18 +1,0 @@
-export class Plugin {
-  constructor() {
-    this.plugins = {};
-  }
-  add(name, plugin) {
-    if (name) {
-      this.plugins[name] = plugin;
-    }
-  }
-  remove(name) {
-    if (name) {
-      delete this.plugins[name];
-    }
-  }
-  getAll() {
-    return this.plugins;
-  }
-}
